requirements.txt
pydantic
typing
langchain
openai
langchain_community
langchain-openai
python-dotenv

